opt glu

+++ include/type.h
@@ -1,4 +1,4 @@
 #ifndef TYPE_H
 #define TYPE_H
-typedef float REAL;
+typedef double REAL;
 #endif
